moved optimized image field to a shared directory
